// src/types/company.ts
export interface CompanyProfile {
  logo_url: string;
  signature_url: string;
  company_name: string;
  address: string;
  phone: string;
  email: string;
  gstin: string;
  state: string;
  bank_name: string;
  account_number: string;
  ifsc_code: string;
  account_holder_name: string;
  user_id?: string;
}